Rebuild brainfuck ram system
I decided this would be a lot easier to follow (and show more than 30
ram slots) if this was at the top of the screen and highlighted what was
currently there.

There's no need for a trail when it comes to ram access given you can't
move arbitrarily around in the ram stack

I removed the ability for memory access to wrap because that doesn't
seem to be a requirement and it makes everything a hell of a lot easier
with regard to creating new things on the stack if you can't make 300k
entries in one go with a single `<`.